delete unused code

+++ Audio.py
@@ -214,7 +214,7 @@ class AudioUIApp(QtWidgets.QMainWindow, AudioUI.Ui_MainWindow):
             # self.data_audio_file = AudioUIApp.butter_lowpass_filter(data=self.data_audio_file,
                                     # cutoff=6000, sample_rate=self.target_sample_rate, order=5)
 
-        self.data_audio_file = self.data_audio_file.astype(np.int16)
+        # self.data_audio_file = self.data_audio_file.astype(np.int16)
 
     
     @staticmethod
